+++ src/ChildPCGuard.AdminPanel/Pages/LogsPage.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace ChildPCGuard.AdminPanel.Pages;
+
+public partial class LogsPage : Page
+{
+    public LogsPage()
+    {
+        InitializeComponent();
+    }
+}